updated files based on updates from story XAPTC-7
