+++ services/xmlWriter.go
@@ -1,4 +1,4 @@
-package main
+package services
 
 import (
 "encoding/xml"
